style: add ipython dependency and move interactive.py to utils
