+++ pyproject.toml
@@ -33,6 +33,7 @@ dependencies = [
     "hydra-core>=1.3.2",
     "omegaconf>=2.3.0",
     "lingua-language-detector>=2.1.0",
+    "ipython>=9.2.0",
 ]
 
 [project.urls]
