+++ src/job_post_nlp/evaluate.py
@@ -10,7 +10,7 @@ from dvclive import Live
 from matplotlib.figure import Figure
 from spacy.tokens import DocBin
 
-from job_post_nlp.interactive import try_inter
+from job_post_nlp.utils.interactive import try_inter
 
 try_inter()
 from job_post_nlp.prepare import corpus_unpack, register_extensions  # noqa: E402
