+++ src/job_post_nlp/utils/interactive.py
